toolset/pyproject.toml
[project]
name = "toolset"
version = "0.1.0"
description = "HuMCP Server - Tool execution service exposing MCP and REST APIs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "fastmcp>=3.0.0rc2",
    "pandas>=2.3.3",
    "duckdb>=1.4.2",
    "tavily-python>=0.7.13",
    "markitdown[all]>=0.1.4",
    "google-api-python-client>=2.168.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "apscheduler>=3.11.2",
    "httpx>=0.28.1",
    "restrictedpython>=8.1",
    "pytesseract>=0.3.13",
    "pillow>=12.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "minio>=7.2.0",
    "openinference-instrumentation-anthropic>=0.1.20",
    "python-jose[cryptography]>=3.3.0",
    "shared",
    # search
    "duckduckgo-search>=7.0.0",
    "baidusearch>=1.0.3",
    "exa-py>=1.0.0",
    "google-search-results>=2.4.0",
    "seltz>=0.1.0",
    "valyu>=0.1.0",
    # web_scraping
    "crawl4ai>=0.4.0",
    "firecrawl-py>=1.0.0",
    "spider-client>=0.1.0",
    "newspaper4k>=0.9.0",
    "lxml-html-clean>=0.4.0",
    "trafilatura>=2.0.0",
    "agentql>=0.3.0",
    "browserbase>=1.0.0",
    "scrapegraph-py>=1.0.0",
    "linkup-sdk>=0.2.0",
    "apify-client>=1.8.0",
    # messaging
    "slack-sdk>=3.33.0",
    "resend>=2.0.0",
    "twilio>=9.0.0",
    # project_management
    "jira>=3.8.0",
    "atlassian-python-api>=3.41.0",
    "todoist-api-python>=2.1.0",
    # finance
    "yfinance>=0.2.0",
    # openbb — excluded due to ruff/uvicorn version conflicts; lazy-imported in finance/openbb.py
    # social
    "praw>=7.8.0",
    "authlib>=1.3.0",
    "youtube-transcript-api>=0.6.0",
    # research
    "arxiv>=2.1.0",
    "pypdf>=5.0.0",
    "wikipedia>=1.4.0",
    # memory
    "mem0ai>=0.1.0",
    "zep-cloud>=2.0.0",
    # media
    "openai>=1.0.0",
    "replicate>=1.0.0",
    "lumaai>=1.0.0",
    "fal-client>=0.5.0",
    # moviepy — excluded due to pillow<12 conflict; lazy-imported in media/moviepy_video.py
    "opencv-python>=4.10.0",
    # audio
    "elevenlabs>=1.0.0",
    "cartesia>=1.0.0",
    # calendar (no extra deps — httpx only)
    # cloud
    "boto3>=1.35.0",
    "docker>=7.0.0",
    "e2b-code-interpreter>=1.0.0",
    # database
    "neo4j>=5.0.0",
    "redshift-connector>=2.1.0",
    # google
    "google-cloud-bigquery>=3.25.0",
]

[tool.uv.sources]
shared = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# Project-specific dev dependencies (inherits shared deps from root)
[dependency-groups]
dev = [
    "grafi>=0.0.34",
    "types-pyyaml>=6.0.12.20250915",
    "pytest-httpx>=0.36.0",
]

[[tool.mypy.overrides]]
module = "yaml.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
testpaths = ["tests", "tests_integration"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::RuntimeWarning:pydub",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = ["src/tools/google/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true

[tool.mypy]
python_version = "3.13"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
check_untyped_defs = false
